refactor(api): manage the CORS in the shared module
