+++ api/subscriptions.js
@@ -16,7 +16,7 @@ const isEmail = require('validator/lib/isEmail')
 const { decamelizeKeys } = require('humps')
 const { default: isMobilePhone } = require('validator/lib/isMobilePhone')
 const { CMS_ENDPOINT_DEPRECATED } = require('../configs/config')
-const corsMiddle = require('./middle/cors-middle-express')
+const { handleExpressCors: handleCors } = require('./middle/cors.js')
 
 const apiHost = CMS_ENDPOINT_DEPRECATED
 
@@ -74,7 +74,7 @@ const setCommonValue = (req, res, next) => {
 router.use(bodyParser.json())
 
 // For CORS non-simple requests
-router.options('/*', corsMiddle, (res) => {
+router.options('/*', handleCors, (res) => {
   res.header('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE,OPTIONS')
   res.header(
     'Access-Control-Allow-Headers',
